make acala constants sharable
